refactor: centralize element disabling helper
